pyproject.toml
[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["py4phi*", "cli*"]

[project]
name="py4phi"
requires-python = ">=3.10"
license={file="LICENSE"}
version = "1.0.0"
description="A library for encryption/decryption and analysis of sensitive data."
authors = [
  {name = "Volodymyr Kiriushyn", email = "zalorderon3331@gmail.com" }
]
dependencies = [
    "iniconfig>=2.0.0",
    "packaging>=23.2",
    "pluggy>=1.3.0",
    "py4j>=0.10.9.7",
    "pyspark>=3.5.0",
    "configparser-crypt>=1.1.0",
    "pycryptodome>=3.20.0",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "polars>=0.20.9",
    "click>=8.1.7",
    "scikit-learn>=1.4.1.post1",
    "scipy>=1.12.0",
    "setuptools>=69.1.0"
]

[project.scripts]
py4phi = "cli.cli_main:main"


[project.optional-dependencies]
test = [
    "pytest==7.4.4",
    "pytest-mock==3.12.0",
    "pytest-cov==4.1.0",
    "mypy==1.8.0",
    "ruff==0.1.3"
]

[tool.setuptools]
package-data = {"py4phi" = ["log4j.properties"]}